# docs/api.yml
openapi : 3.0.3
info : 
  title : MyBank API
  version : 2.4.2
  description: |
    Project app for PROG4 cours by : 
    Fandresena (STD22058),
    Sheddy (STD22012),
    Daris (STD21099)
servers:
  - url: https://app.numbank.mg
...
paths:
...
  /balance/{accountid}:
    get:
      tags:
        - "Balance"
      summary: Get all balance about an account
      operationId: getBalancesByAccountId
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: startDateTime
          required: false
          schema:
            type: string
            format: datetime
        - in: query
          name: endDateTime
          required: false
          schema:
            type: string
            format: datetime
      responses:
        200:
          description: Return a list of balance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BalanceHistory'
  /balance/moneydrawal/{accountid}:
    get:
      tags:
        - "Balance"
      summary: Get all money drawal about an account
      operationId: getMoneyDrawalsByAccountId
      parameters:
        - in: path
          name: accountid
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: startDateTime
          required: false
          schema:
            type: string
            format: datetime
        - in: query
          name: endDateTime
          required: false
          schema:
            type: string
            format: datetime
      responses:
        200:
          description: Return a list of money drawal
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MoneyDrawal'
components:
  schemas:
...
    BalanceHistory:
      type: object
      properties:
        id:
          type: number
        balance:
          type: number
          format: double
        dateTime:
          type: string
          format: date
    MoneyDrawal:
      type: object
      properties:
        id:
          type: number
        amount:
          type: number
          format: double
        withDrawalDate:
          type: string
          format: date